Adds some checks

+++ neonwranglerpy/utilities/loadByProduct.py
@@ -1,34 +1,43 @@
 import os.path
 from tempfile import mkdtemp
 from shutil import rmtree
+from re import match, compile
 from neonwranglerpy.utilities.zipsByProduct import zips_by_product
 from neonwranglerpy.utilities.stackByTable import stack_by_table
 
+DATE_PATTERN = compile('20[0-9]{2}-[0-9]{2}')
 
-def load_by_product(dpID,
-                    site='all',
-                    start_date=None,
-                    end_date=None,
-                    package="basic",
-                    release="current",
-                    path='./'):
-    """
-    :param path:
-    :param dpID: productID
-    :param site: site ID
-    :param start_date:
-    :param end_date:
-    :param package:
-    :param release:
-    :return:
-    """
-    if package != "basic" or package != "expanded":
-        print(f"{package} is not a valid package name. Package must be basic or expanded")
-        return
 
-    # TODO: add a check for correct format of product ID
-    # TODO: add a check for AOP data product
-    # TODO: add a check for start and end date
+def load_by_product(
+        dpID,
+        site='all',
+        start_date=None,
+        end_date=None,
+        package="basic",
+        release="current",
+        path='./',
+        save_files=False,
+        stacked_df=False
+):
+    # if package != "basic" or package != "expanded":
+    #     print(f"{package} is not a valid package name. Package must be basic or expanded")
+    #     return
+
+    if not match("DP[1-4]{1}.[0-9]{5}.00[0-9]{1}", dpID):
+        return f"{dpID} is not a properly formatted data product ID. The correct format is DP#.#####.00#, " \
+               f"where the first placeholder must be between 1 and 4."
+
+    if dpID[4:5] == 3 and dpID != "DP1.30012.001":
+        return f'{dpID}, "is a remote sensing data product and cannot be loaded directly to R with this function.Use ' \
+               f'the byFileAOP() or byTileAOP() function to download locally." '
+
+    if len(start_date):
+        if not match(DATE_PATTERN, start_date):
+            return 'startdate and enddate must be either NA or valid dates in the form YYYY-MM'
+
+    if len(end_date):
+        if not match(DATE_PATTERN, end_date):
+            return 'startdate and enddate must be either NA or valid dates in the form YYYY-MM'
 
     args = {
         "dpID": dpID,
@@ -43,9 +52,10 @@ def load_by_product(dpID,
     tempdir = mkdtemp(dir=os.path.dirname(path))
 
     # pass the request to zipsByProduct() to download
-    zips_by_product(dpID, site, start_date, end_date, package, release, path=tempdir)
+    path = zips_by_product(dpID, site, start_date, end_date, package, release, savepath=tempdir)
     # stack the tables by using stackByTable\
-    out = stack_by_table(filepath=tempdir)
+    out = stack_by_table(filepath=path,dpID=dpID,savepath=tempdir,stack_df=stacked_df)
     # removes temp dir
-    rmtree(tempdir)
+    if not save_files:
+        rmtree(tempdir)
     return out
